# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (Admin only)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return the current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by id (Admin or own profile)",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user (Admin only)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User has been deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user (Admin only)",
        "tags": [
          "users"
        ]
      }
    },
...
    "/payments/details": {
      "get": {
        "operationId": "PaymentsController_findAllDetails",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all payment details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentDetailResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all payment details",
        "tags": [
          "payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_createDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDetailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The payment detail has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Donation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a payment detail",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/details/{id}": {
      "get": {
        "operationId": "PaymentsController_findOneDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment Detail ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Payment detail not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a payment detail by ID",
        "tags": [
          "payments"
        ]
      },
      "patch": {
        "operationId": "PaymentsController_updateDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment Detail ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDetailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The payment detail has been updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a payment detail",
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_removeDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment Detail ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The payment detail has been deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a payment detail",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/details/donation/{donationId}": {
      "get": {
        "operationId": "PaymentsController_findByDonation",
        "parameters": [
          {
            "name": "donationId",
            "required": true,
            "in": "path",
            "description": "Donation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Payment detail not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payment details by donation ID",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/create-intent": {
      "post": {
        "operationId": "PaymentsController_createIntent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIntentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The payment intent has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a payment intent",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/confirm-intent/{paymentIntentId}/{donationId}": {
      "post": {
        "operationId": "PaymentsController_confirmIntent",
        "parameters": [
          {
            "name": "paymentIntentId",
            "required": true,
            "in": "path",
            "description": "Payment Intent ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "donationId",
            "required": true,
            "in": "path",
            "description": "Donation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment intent has been confirmed"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Confirm a payment intent",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_handleWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "description": "Stripe webhook signature",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Raw webhook body from Stripe",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          }
        },
        "summary": "Handle payment webhook",
        "tags": [
          "payments"
        ]
      }
    }
  },
  "info": {
    "title": "Masjid Network API",
    "description": "\n# Masjid Network API\n\n## Overview\nThe Masjid Network API provides a comprehensive backend for managing mosques (masjids), fundraising campaigns, donations, and user accounts. This platform connects mosques with their communities and facilitates fundraising for various projects and initiatives.\n\n## Features\n- User authentication and management\n- Masjid/mosque profiles and administration\n- Fundraising campaigns\n- Donation processing\n- Payment handling\n\n## Authentication\nMost endpoints require authentication using JWT Bearer tokens.\n\nTo obtain a token, use the /auth/login endpoint with valid credentials.\nInclude the token in the Authorization header: `Bearer your_token_here`\n\n## Getting Started\n1. Register a user account via /auth/register\n2. Login to obtain a JWT token via /auth/login\n3. Use the token to access protected endpoints\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "firstName": {
            "type": "object",
            "example": "John"
          },
          "lastName": {
            "type": "object",
            "example": "Doe"
          },
          "phone": {
            "type": "object",
            "example": "+1234567890"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MASJID_ADMIN",
              "SUPER_ADMIN"
            ],
            "example": "USER"
          },
          "profileImage": {
            "type": "object",
            "example": "https://example.com/avatar.jpg"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "lastLogin": {
            "type": "object",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "createdAt",
          "updatedAt",
          "isActive"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "profileImage": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          }
        }
      },
      "AdminUpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "profileImage": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN",
              "MASJID_ADMIN",
              "SUPER_ADMIN"
            ],
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "PaymentDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the payment details"
          },
          "provider": {
            "type": "string",
            "description": "Payment provider (e.g. STRIPE, PAYPAL)"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Payment method ID from provider"
          },
          "receiptUrl": {
            "type": "string",
            "description": "URL to receipt/invoice"
          },
          "metadata": {
            "type": "object",
            "description": "Additional provider-specific data"
          },
          "donationId": {
            "type": "string",
            "description": "ID of the related donation"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment details were created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the payment details were last updated"
          }
        },
        "required": [
          "id",
          "provider",
          "paymentMethodId",
          "receiptUrl",
          "metadata",
          "donationId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePaymentDetailDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "Payment provider (e.g. STRIPE, PAYPAL)"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Payment method ID from provider"
          },
          "receiptUrl": {
            "type": "string",
            "description": "URL to receipt/invoice"
          },
          "metadata": {
            "type": "object",
            "description": "Additional provider-specific data"
          },
          "donationId": {
            "type": "string",
            "description": "ID of the related donation"
          }
        },
        "required": [
          "provider",
          "paymentMethodId",
          "receiptUrl",
          "metadata",
          "donationId"
        ]
      },
      "UpdatePaymentDetailDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "Payment provider (e.g. STRIPE, PAYPAL)"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Payment method ID from provider"
          },
          "receiptUrl": {
            "type": "string",
            "description": "URL to receipt/invoice"
          },
          "metadata": {
            "type": "object",
            "description": "Additional provider-specific data"
          },
          "donationId": {
            "type": "string",
            "description": "ID of the related donation"
          }
        }
      },
      "CreatePaymentIntentDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Amount to process (in cents)"
          },
          "currency": {
            "type": "string",
            "description": "Currency code (e.g. USD)"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method to use"
          },
          "campaignId": {
            "type": "string",
            "description": "ID of the campaign to donate to"
          },
          "isAnonymous": {
            "type": "boolean",
            "description": "Whether the donation should be anonymous"
          },
          "message": {
            "type": "string",
            "description": "Optional message to include with donation"
          }
        },
        "required": [
          "amount",
          "currency",
          "paymentMethod",
          "campaignId",
          "isAnonymous",
          "message"
        ]
      },
      "PaymentIntentResponseDto": {
        "type": "object",
        "properties": {
          "clientSecret": {
            "type": "string",
            "description": "Client secret for Stripe payment intent"
          },
          "paymentIntentId": {
            "type": "string",
            "description": "Payment intent ID"
          },
          "donationId": {
            "type": "string",
            "description": "Temporary donation ID"
          }
        },
        "required": [
          "clientSecret",
          "paymentIntentId",
          "donationId"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      }
    }
  }
}